+++ package.json
@@ -8,8 +8,8 @@
     "build-storybook": "build-storybook -s public"
   },
   "dependencies": {
-    "@intellective/core": "2020.1.1-alpha.89",
-    "@intellective/unity-api-mocks": "^1.0.0",
+    "@intellective/core": "2020.1.1",
+    "@intellective/unity-api-mocks": "^1.0.5",
     "@material-ui/core": "^4.11.0",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/lab": "^4.0.0-alpha.56",
