+++ server/rest.js
@@ -8,11 +8,11 @@ const config = R.mergeDeepRight(unityApiMocks.defaultConfig, domainConfig);
 
 const usersApi = require("./api/usersApi");
 
-const userTransformer = record => R.mergeDeepRight(record, {_links: {'custom.verify': {href: '/api/users/verify'}}});
+const useCustomUserLinks = record => R.mergeDeepRight(record, {_links: {'custom.verify': {href: '/api/users/verify'}}});
 
 module.exports = function (app) {
 	unityApiMocks.configApi(app, config);
-	unityApiMocks.usersApi(app, config, [], userTransformer);
+	unityApiMocks.usersApi(app, config, [], useCustomUserLinks);
 	unityApiMocks.contentApi(app, config);
 	unityApiMocks.casesApi(app, config);
 	unityApiMocks.foldersApi(app, config);
