+++ src/stories/GridView/GridViewToolbar.stories.js
@@ -4,60 +4,56 @@ import {useDispatch, useSelector} from "react-redux";
 
 import * as R from "ramda";
 
-import Alert from "@material-ui/lab/Alert";
 import IconButton from "@material-ui/core/IconButton";
 import Tooltip from "@material-ui/core/Tooltip";
 import Typography from "@material-ui/core/Typography";
-import TableCell from "@material-ui/core/TableCell";
 import Toolbar from "@material-ui/core/Toolbar";
 import LinearProgress from "@material-ui/core/LinearProgress";
 import withStyles from "@material-ui/styles/withStyles";
-
-import GetAppIcon from '@material-ui/icons/GetApp';
 import Print from "@material-ui/icons/Print";
 
 import {
+	DefaultActionFactory,
+	DefaultActionMapper,
 	DefaultComponentFactory,
 	DefaultFormSubmitHandlerMapper,
 	defaultGridViewSettings,
+	DefaultViewForm,
+	FormField,
+	forms,
 	GridView,
+	parseFormData,
+	resources,
+	submitRequest,
+	TwoColumnsLayout,
+	withActionBackdrop,
+	withActionView,
+	withFormSubmitAction,
 	withGridViewActionExecutor,
 	withGridViewConfigLoader,
+	withGridViewDataLoader,
 	withGridViewDefaultActions,
 	withGridViewModelBulkActions,
 	withGridViewPagination,
-	withGridViewDataLoader,
 	withGridViewSelection,
 	withGridViewSettings,
-	withGridViewSorting,
-	withGridViewRowActions,
-	useDefaultColumnRenderer,
-	grids,
-	FormField,
-	DefaultActionFactory,
-	DefaultActionMapper,
-	withActionView,
-	withActionBackdrop,
-	withFormSubmitAction,
-	DefaultViewForm,
+	withGridViewSorting
 } from "@intellective/core";
 
 import AppPage from "../../components/AppPage/AppPage";
-import {isDocument} from "@intellective/core/build/factories/resources";
-import {getLinkByRel, parseFormData, submitRequest} from "@intellective/core/build/actions";
-import {defaultConfig as actionFactoryConfig} from "@intellective/core/build/factories/DefaultActionFactory";
-import {
-	clearFormState,
-	updateFieldValue,
-	updateFormState
-} from "@intellective/core/build/store/actions/forms";
 
 export default {title: 'Examples/Grid View'};
+
+const getLinkByRel = (_links, type, name) => _links && (_links[name] || _links[type] || {}).href;
+
+const {isDocument} = resources;
+
+const {clearFormState, updateFieldValue, updateFormState} = forms;
 
 /*
 * Add custom toolbar action to the grid component
 */
-export const UsingToolbarAction = () => {
+export const UsingCustomToolbarAction = () => {
 
 	/**
 	 * Custom action that prints "Hello World" message below the grid
@@ -132,7 +128,9 @@ export const UsingToolbarAction = () => {
  */
 const withFormCleanup = R.curry((WrappedComponent, props) => {
 	const {formId} = props;
+
 	const dispatch = useDispatch();
+
 	const cleanup = () => dispatch(clearFormState(formId));
 
 	useEffect(() => {
@@ -145,7 +143,7 @@ const withFormCleanup = R.curry((WrappedComponent, props) => {
 /*
 * Add toolbar bulk action with configured view to the grid component
 */
-export const UsingToolbarBulkActionWithConfiguredView = () => {
+export const UsingCustomBulkAction = () => {
 	/**
 	 * Condition to detect Verify action
 	 */
@@ -210,22 +208,31 @@ export const UsingToolbarBulkActionWithConfiguredView = () => {
 		])(action);
 	});
 
+	const settings = {
+		variant: 'dialog',
+		fullScreen: true,
+		maxWidth: 'xl',
+		innerMaxWidth: 'lg',
+		margin: 'dense',
+		Layout: TwoColumnsLayout
+	};
+
 	/**
 	 * Action factory component with redefined ActionMapper and FormSubmitHandlerMapper components
 	 */
 	const ActionFactory =
-		new DefaultActionFactory(actionFactoryConfig, DomainActionMapper, DomainFormSubmitHandlerMapper);
+		new DefaultActionFactory(settings, DomainActionMapper, DomainFormSubmitHandlerMapper);
 
 	return (
 		<AppPage href="/api/1.0.0/config/perspectives/search/dashboards/page14"
-				 ActionFactory={ActionFactory}/>
+		         ActionFactory={ActionFactory}/>
 	);
 };
 
 /*
 * Add custom toolbar bulk action with custom view to the grid component
 */
-export const UsingToolbarBulkActionWithCustomView = () => {
+export const UsingCustomBulkActionWithCustomView = () => {
 	/**
 	 * Condition to detect Verify action
 	 */
@@ -282,8 +289,8 @@ export const UsingToolbarBulkActionWithCustomView = () => {
 					</Typography>
 					<form id={formId} noValidate role="form" aria-label="Details View">
 						<FieldComponent {...R.prop('age', formData)} formatValue={x => `${x} years`}
-										data={[{value: 0}, {value: 18}, {value: 25}, {value: 35}, {value: 45}, {value: 60}, {value: 99}]}
-										onChange={(value) => dispatch(updateFieldValue(formId, 'age', value))}/>
+						                data={[{value: 0}, {value: 18}, {value: 25}, {value: 35}, {value: 45}, {value: 60}, {value: 99}]}
+						                onChange={(value) => dispatch(updateFieldValue(formId, 'age', value))}/>
 					</form>
 					<Typography display='block' variant="body2" color='textSecondary' mt={3}>
 						<p><sup>1</sup>A business must identify and understand its target audience if their marketing campaign is
@@ -339,241 +346,22 @@ export const UsingToolbarBulkActionWithCustomView = () => {
 		])(action);
 	});
 
+	const settings = {
+		variant: 'dialog',
+		fullScreen: true,
+		maxWidth: 'xl',
+		innerMaxWidth: 'lg',
+		margin: 'dense',
+		Layout: TwoColumnsLayout
+	};
+
 	/**
 	 * Action factory component with redefined ActionMapper and FormSubmitHandlerMapper components
 	 */
-	const ActionFactory =
-		new DefaultActionFactory(actionFactoryConfig, DomainActionMapper, DomainFormSubmitHandlerMapper);
+	const ActionFactory = new DefaultActionFactory(settings, DomainActionMapper, DomainFormSubmitHandlerMapper);
 
 	return (
 		<AppPage href="/api/1.0.0/config/perspectives/search/dashboards/page14"
-				 ActionFactory={ActionFactory}/>
-	);
-};
-
-/*
-* Add custom column renderer
-*/
-export const UsingColumnRendering = () => {
-
-	const mapValue = R.cond([
-		[R.is(Array), R.map(R.when(R.is(Object), R.prop('value')))],
-		[R.is(Object), R.prop('value')],
-		[R.T, R.identity]
-	]);
-
-	const renderStatus = R.curry((column, value) => {
-		const {name} = column;
-
-		if (value === undefined || value === null || value === '') {
-			return null;
-		}
-
-		const isCaseStatus = "case_status" === name;
-
-		const severity = R.cond([
-			[R.equals("Open"), R.always("info")],
-			[R.equals("Error"), R.always("error")],
-			[R.equals("Close"), R.always("success")],
-			[R.T, R.always("warning")],
-		])(value);
-
-		const icon = isCaseStatus ? {icon: false} : {};
-
-		return <Alert variant={isCaseStatus ? "filled" : "outlined"} severity={severity} {...icon}>{value}</Alert>;
-	});
-
-	const domainStatusColumnRenderer = (value, row, column) => R.compose(renderStatus(column), mapValue)(value);
-
-	const withDomainColumnRendering = R.curry((WrappedGrid, props) => {
-		const {columns = []} = props;
-
-		const isCaseStatusColumn = R.anyPass([R.propEq('name', 'case_status'), R.propEq('name', 'task_status')]);
-
-		const customStatusColumn = R.when(isCaseStatusColumn, column => ({
-			...column,
-			renderer: domainStatusColumnRenderer
-		}));
-
-		const _columns = R.map(customStatusColumn, columns);
-
-		return <WrappedGrid {...props} columns={_columns}/>;
-	});
-
-	/**
-	 * Custom Grid View factory with the custom column renderer addition
-	 */
-	const GridViewFactory = (props) => {
-		const ComposedGridView = R.compose(
-			withGridViewConfigLoader,
-			withGridViewSettings(defaultGridViewSettings),
-			withGridViewDefaultActions,
-			withGridViewModelBulkActions,
-			withGridViewActionExecutor,
-			withGridViewDataLoader,
-			withGridViewSorting,
-			withGridViewSelection,
-			withGridViewPagination,
-			withDomainColumnRendering
-		)(GridView);
-
-		return (
-			<ComposedGridView {...props}/>
-		);
-	};
-
-	const DomainComponentMapping = R.cond([
-		[R.propEq('type', 'grid'), GridViewFactory],
-	]);
-
-	/**
-	 *  Customize the default component factory logic with simple boolean condition so that the custom component factory comes first
-	 */
-	const DomainComponentFactory = (props) => DomainComponentMapping(props) || DefaultComponentFactory(props);
-
-	return (
-		<AppPage href="/api/1.0.0/config/perspectives/search/dashboards/page12"
-		         ComponentFactory={DomainComponentFactory}/>
-	);
-};
-
-/*
-* Add custom column action
-*/
-export const UsingCustomColumnAction = () => {
-
-	/**
-	 * Custom action that invokes download action on fullName column click instead of view action
-	 */
-	const withCustomColumnActionType = R.curry((WrappedGrid, props) => {
-
-		const mappedColumnName = 'gender';
-
-		const useColumnActionType = R.cond([[R.propEq('name', mappedColumnName), R.always('download')]]);
-
-		const customActionRenderer = R.curry((column) => {
-
-			return (value, record, column) => {
-				const handleClick = (event) => {
-					event.preventDefault();
-
-					column.onClick && column.onClick(record);
-				};
-
-				return <Tooltip title={column.label} role="tooltip">
-					<GetAppIcon cursor='pointer' color="inherit" onClick={handleClick}/>
-				</Tooltip>;
-			};
-		});
-
-		// eslint-disable-next-line react-hooks/rules-of-hooks
-		const useColumnRenderer = column => R.cond([[R.propEq('name', mappedColumnName), customActionRenderer]])(column) || useDefaultColumnRenderer(column);
-
-		return <WrappedGrid {...props} useColumnActionType={useColumnActionType} useColumnRenderer={useColumnRenderer} />;
-	});
-
-	/**
-	 * Custom Grid View factory with the custom column action
-	 */
-	const GridViewFactory = (props) => {
-		const ComposedGridView = R.compose(
-			withCustomColumnActionType,
-			withGridViewConfigLoader,
-			withGridViewSettings(defaultGridViewSettings),
-			withGridViewDefaultActions,
-			withGridViewModelBulkActions,
-			withGridViewRowActions,
-			withGridViewActionExecutor,
-			withGridViewDataLoader,
-			withGridViewSorting,
-			withGridViewSelection,
-			withGridViewPagination,
-		)(GridView);
-
-		return (
-			<ComposedGridView {...props}/>
-		);
-	};
-
-	const DomainComponentMapping = R.cond([
-		[R.propEq('type', 'grid'), GridViewFactory],
-	]);
-
-	/**
-	 *  Customize the default component factory logic with simple boolean condition so that the custom component factory comes first
-	 */
-	const DomainComponentFactory = (props) => DomainComponentMapping(props) || DefaultComponentFactory(props);
-
-	return (
-		<AppPage href="/api/1.0.0/config/perspectives/search/dashboards/page1"
-				 ComponentFactory={DomainComponentFactory}/>
-	);
-};
-
-/*
-* Add row double-click handler
-*/
-
-/**
- *  Custom action that implements customCellRenderer with onDoubleClick handler
- */
-const findCustomAction = actionName => R.find(R.propEq("name", actionName));
-
-const withUseCustomCellRenderer = R.curry((WrappedGrid, props) => {
-	const {id, actions} = props;
-	const dispatch = useDispatch();
-
-	const action = findCustomAction('view')(actions);
-
-	const useCustomCellRenderer = R.curry((classes, row, column) => {
-
-		const value = row[column.name];
-
-		return (
-			<TableCell onDoubleClick={() => dispatch(grids.updateGridCurrentAction(id, {action, selected: row}))} key={`${row.id}-${column.name}`} className={classes.tableCell} scope="row" variant="body">
-				{column.renderer ? column.renderer(value, row, column) : value}
-			</TableCell>
-		);
-	});
-
-	return <WrappedGrid {...props} useCellRenderer={useCustomCellRenderer}/>;
-});
-
-export const UsingDoubleClickHandler = () => {
-
-	/**
-	 * Custom Grid View factory with double click handler addition
-	 */
-	const GridViewFactory = (props) => {
-		const ComposedGridView = R.compose(
-			withGridViewConfigLoader,
-			withGridViewSettings(defaultGridViewSettings),
-			withGridViewDefaultActions,
-			withGridViewModelBulkActions,
-			withGridViewActionExecutor,
-			withGridViewDataLoader,
-			withGridViewSorting,
-			withGridViewSelection,
-			withGridViewPagination,
-			withUseCustomCellRenderer
-		)(GridView);
-
-		return (
-			<ComposedGridView {...props}/>
-		);
-	};
-
-	const DomainComponentMapping = R.cond([
-		[R.propEq('type', 'grid'), GridViewFactory],
-	]);
-
-	/**
-	 *  Customize the default component factory logic with simple boolean condition so that the custom component factory comes first
-	 */
-	const DomainComponentFactory = (props) => DomainComponentMapping(props) || DefaultComponentFactory(props);
-
-	return (
-		<AppPage href="/api/1.0.0/config/perspectives/search/dashboards/page1"
-				 ComponentFactory={DomainComponentFactory}/>
+		         ActionFactory={ActionFactory}/>
 	);
 };
